requirements.txt
Django
django-allauth
djangorestframework
celery
redis
django-crispy-forms
crispy-tailwind
django-modeltranslation
Pillow
python-decouple
django-cors-headers
gunicorn
whitenoise
psycopg2-binary
boto3
django-storages
twilio
stripe
sslcommerz-lib
reportlab
pandas
openpyxl
django-ratelimit
django-debug-toolbar
python-dotenv
requests
msal
google-auth
google-auth-oauthlib
google-auth-httplib2
zoom
django-extensions